requirements.txt
#pip install -r requirements.txt
uvicorn
fastapi
SQLAlchemy
passlib
pyjwt
bcrypt
python-multipart